+++ src/main/java/com/example/demo/repositories/CheckInRepository.java
@@ -6,11 +6,11 @@ import org.springframework.lang.Nullable;
 
 import java.util.List;
 
-public interface CheckInRepository extends JpaRepository<CheckIn, Long> {
+public interface CheckInRepository extends JpaRepository<CheckIn, Integer> {
     //delete
-    List<CheckIn> findByClientId(Long clientId);
+    List<CheckIn> findByClientId(int clientId);
 
-    List<CheckIn> findByRoomId(Long clubId);
+    List<CheckIn> findByRoomId(int clubId);
 
     <S extends CheckIn> S save(@Nullable S coach);
 }
